Help Menu & Large Number Math
Fixed the math operations to be capable of working with numbers larger
than 3 digits. When captured by voice, numbers of length 4 or more were
accruing comma seperators which screwed up integer calaculations. After
being tokenized, the strings are now stripped of commas and parsed into
integers.

Added a new command called HELP. When a user says help, the app will
explain that it expects a number in reference to the numbered list of
commands displayed in the command dialog box. After a number is heard,
the app will give further explanation of how to use the command and/or
what the result should be.