Fix for help trigger
When a user asks for help, the application starts listening for an
integer number in subsequent commands. Previously, the listener
persisted even if the user decided not to suplly the program with an
integer. Now, if a user says anything besides a number after asking for
help, the special listener for integers is dismissed.