Fix crash from unexpected operator
When the application hears a number, it begins to expect a math
question. If the full request can be divided into 3 arguments, the
first number, an operator, and a second number, the app solves the
expression. Supplying an unknown operator previously caused a
malfunction. Now the application has a response to unrecognized
operators.